# src/static/swagger.yaml
swagger: '2.0'
info:
  title: Library API
  description: Library API - EDC API Workshop
  version: 1.0.0
  contact:
    email: oyron@equinor.com
host: 'localhost:3100'
basePath: /api
schemes:
  - http
tags:
  - name: library
    description: A library of books
paths:
  /books:
    get:
      tags:
        - library
      summary: Return all books
      operationId: getBooks
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Book'
    post:
      tags:
        - library
      summary: Add a new book to the library
      operationId: addBook
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: bookData object that needs to be added to the library
          required: true
          schema:
            $ref: '#/definitions/BookData'
      responses:
        '201':
          description: Created
          schema:
            items:
              $ref: '#/definitions/Book'
  '/books/{id}':
    get:
      tags:
        - library
      summary: Get book
      operationId: getBook
      produces:
        - application/json
        - text/html
      parameters:
        - name: id
          in: path
          description: Book id
          type: integer
          required: true
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/Book'
        '400':
          description: Bad request
        '404':
          description: Not found
    put:
      tags:
        - library
      summary: Update existing book
      operationId: updateBook
      produces:
        - application/json
        - text/html
      parameters:
        - name: id
          in: path
          description: Book id
          type: integer
          required: true
        - name: body
          in: body
          description: book data to be updated in library
          required: true
          schema:
            $ref: '#/definitions/BookData'
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/Book'
        '400':
          description: Bad request
        '404':
          description: Not found
    delete:
      tags:
        - library
      summary: Delete existing book
      operationId: deleteBook
      produces:
        - text/html
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Book id
      responses:
        '204':
          description: No content
        '400':
          description: Bad request
        '404':
          description: Not found
definitions:
  Book:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      author:
        type: string
  BookData:
    type: object
    properties:
      title:
        type: string
      author:
        type: string
externalDocs:
  description: API documentation
  url: 'http://localhost:3000'